# blog/docs/swagger.yaml
openapi: 3.0.3
info:
  description: "This is server for blog-platform server."
  version: "1.0.11"
  title: "Swagger blog-platform"
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/user/{user_id}/create-post:
    post:
      tags:
        - posts
      summary: "Create New Post"
      description: "Create New Post"
      operationId: "CreatePost"
      parameters:
        - name: user_id
          in: path
          description: id of the particular user
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePost'
          application/xml:
            schema:
              $ref: '#/components/schemas/CreatePost'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreatePost'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '405':
          description: Invalid input
  /api/user/{user_id}/post/{post_id}:
    get:
      tags:
        - posts
      summary: Find post by ID
      description: Returns a single post
      operationId: getPostById
      parameters:
        - name: user_id
          in: path
          description: ID of particular user
          required: true
          schema:
            type: integer
            format: int64
        - name: post_id
          in: path
          description: tag id that need to be return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
            application/xml:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Invalid ID supplied
        '404':
          description: post not found
    put:
      tags:
        - posts
      summary: Updates a post in the store with form data
      description: ''
      operationId: updatePostWithForm
      parameters:
        - name: user_id
          in: path
          description: ID of particular user
          required: true
          schema:
            type: integer
            format: int64
        - name: post_id
          in: path
          description: ID of post that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePost'
          application/xml:
            schema:
              $ref: '#/components/schemas/CreatePost'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreatePost'
      responses:
        '200':
          description: successful post update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Bad Request

    delete:
      tags:
        - posts
      summary: Deletes a post
      description: delete a post
      operationId: deletePost
      parameters:
        - name: user_id
          in: path
          description: ID of particular user
          required: true
          schema:
            type: integer
            format: int64
        - name: post_id
          in: path
          description: post id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: No Content
        '400':
          description: Invalid post value
  /api/posts:
    get:
      tags:
        - posts
      summary: list all posts
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/to"
        - $ref: "#/components/parameters/from"
      responses:
        '200':
          description: list all posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '500':
          description: internal server error
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: theUser
        created_at:
          type: string
          format: date-time
          example: 2022-08-14T19:33:16.428870284+05:30
        updated_at:
          type: string
          format: date-time
          example: 2022-08-14T19:33:16.428870284+05:30
      xml:
        name: user
    CreatePost:
      type: object
      properties:
        title:
          type: string
          example: "theUser"
        Content:
          type: string
          example: "content"
        tags_id:
          type: integer
          example: 1
      xml:
        name: posts
...
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          example: "classic"
        content:
          type: string
          example: "test blog"
        author:
          $ref: '#/components/schemas/User'
        tags_id:
          type: integer
          example: 1
        author_id:
          type: integer
          example: 1
        tags:
          type: array
          items:
            type: string
        comments:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
          example: 2022-08-14T19:33:16.428870284+05:30
        updated_at:
          type: string
          format: date-time
          example: 2022-08-14T19:33:16.428870284+05:30
...
  parameters:
    limit:
      name: limit
      in: query
      description: Number of records to return
      schema:
        type: integer
        format: int64
    to:
      name: to
      in: query
      description: first index of the page
      schema:
        type: integer
        format: int64
    from:
      name: from
      in: query
      description: last index of the previous page
      schema:
        type: integer
        format: int64